use two different values for "outside" and "border"
This commit changes the type of the heightmap from uint8_t to int16_t, using a
typedef so that changing it again is no problem.

That change was motivated by the need to use a different value for outside and
border cells. Originally, -1 and 0 were chosen, but that breaks the divergence
calculation (which was wrongly assuming that outside cells were valid).

In order to keep that calculation simple, now:
 * outside cells have a value of 0
 * border cells have a value of 1

Note that a (very good) side-effect of this commit is that now border cells are
rightly considered in the divergence calculation. Only outside cells are
ignored.

+++ test/test_dstar_multipath.cc
@@ -140,7 +140,7 @@ BOOST_AUTO_TEST_CASE( test_multipath_forest )
         ofstream out("multipath_forest_skeleton.txt");
         SkeletonExtraction skeleton(Size, Size);
         Skeleton skeleton_points = skeleton.processEdgeSet(result.first, result.second);
-        vector<uint8_t> map = skeleton.getHeightmap();
+        vector<int16_t> map = skeleton.getHeightmap();
 
         out << Size << " " << Size << "\n";
         for (int y = 0; y < Size; ++y)
